requirements.txt
pygame
scikit-learn
seaborn
numpy
scipy
opencv-python
matplotli
pandas